db_settings_to_db_string: Support more URL formats
All or any of username, password, host or port could be empty.  Make
sure that the URL is generated correctly then.  Also, there is no need
to specify the default port: If port is not specified in the settings,
just leave it out of the URL.

Here's a couple examples of valid URLs that were not supported before:

  * postgresql:///dbname

      - Local PostgreSQL database named "dbname", connecting via unix
        socket without any username or password.

  * mysql://localhost/dbname

      - MySQL database "dbname" on localhost