Use yaml.safe_load instead of yaml.load
PyYAML has deprecated yaml.load() when a Loader parameter is not
specified. This project uses yaml.load() without specifying a Loader
parameter, which will fail for users with PyYAML above version 5.1.

PyYAML has always offered safe_load() as an alternative so hopefully
that will work for the purposes of this project.

Please see the following page for more information:
  https://github.com/yaml/pyyaml/wiki/PyYAML-yaml.load(input)-Deprecation